(Hopefully) Fix miscomputing isomorphisms bug, first appearing in M_{1,4}.
The problems was lying in `Fatgraph._neighbors` and
`Fatgraph._extend_map` that did not check whether the *destination*
vertex was already in the image of the (partial) `pv` permutation
(they did the check for the source vertex, tho).  This resulted in the
same vertex being added twice, and `pv` losing its property of being
bijective.

(Actually, the `Permutation` class should enforce this constraint and
not allow extensions that violate the bjectivity invariant property.)

  modified:   combinatorics.py
  modified:   graph_homology.py
  modified:   main.py
  modified:   rg.py

git-svn-id: https://fatghol.googlecode.com/svn/trunk@609 e2c0af65-7b97-2394-be7d-76dd08efcea0